Add files via upload
ball-drop-final contains code to create a kronecker graph via the generalized ball dropping method. In this method, a ball is dropped through each layer of the kronecker network with a weighted probability distribution until it reaches the final layer at which point an edge is created. Balls are continually dropped until the expected number of edges is created.