Add files via upload
Code to generate a random Erdos-Renyi graph with a fixed number of nodes and edges by utilizing 2 iterations of a generalized unranking procedure. The unranking procedure in this code maps a linear index back to a user-defined length permutation with values up to a user-fixed maximum. All unranking is done using lexicographic ordering.